refactor: 카드포멧 커스텀 훅 유틸 함수로 변경
